requirements.txt
python-dotenv
google-generativeai
langchain
PyPDF2
faiss-cpu
streamlit
-e .
langchain-google-genai>=2.1.12
google-generativeai>=0.8.0
langchain-core>=0.3.0
streamlit
faiss-cpu
python-dotenv
langchain-community

streamlit>=1.28.0
langchain-google-genai>=2.1.12
google-generativeai>=0.8.0
langchain-core>=0.3.0
langchain-community>=0.3.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
python-dotenv>=1.0.0
langchain>=0.3.0

# Text-to-Speech dependencies
gtts>=2.3.0
pygame>=2.5.0

# Additional audio processing (optional)
pydub>=0.25.1